S03P23A205-98 backEdit

+++ backend/accounts/models.py
@@ -7,7 +7,7 @@ class CustomUser(AbstractUser):
 	email = models.EmailField(unique=True, blank=False)
 	play = models.IntegerField(default=0) # AI 대전 횟수
 	win = models.IntegerField(default=0) # AI 대전 승
-	rate = models.FloatField(default=0) # AI 대전 승/ 횟수
+	rate = models.IntegerField(default=0) # AI 대전 승/ 횟수
 	score = models.IntegerField(default=0) # 묘수풀이 점수
 
 	USERNAME_FIELD = 'email'
